Added new fields in Person model (LifeIndicators)
