+++ src/game/models/equipment.go
@@ -1,7 +1,5 @@
 package models
 
-import "fmt"
-
 type Equipmenter interface {
 	Equip(*Item)
 }
@@ -20,25 +18,16 @@ type HandWeapon struct {
 	Item *Item
 }
 
-func (h *HandWeapon) Equip(item *Item) {
-	h.Item = item
-}
-
 type BodyArmor struct {
 	Item *Item
 }
 
+func (h *HandWeapon) Equip(item *Item) {
+	h.Item = item
+}
+
 func (b *BodyArmor) Equip(item *Item) {
 	b.Item = item
-}
-
-type Item struct {
-	Title      string
-	Properties Properties
-}
-
-func (i Item) String() string {
-	return fmt.Sprintf("%q with props: %+v", i.Title, i.Properties)
 }
 
 func CreateEmptyEquipment() *Equipment {
@@ -57,28 +46,4 @@ func (e *Equipment) GetEquipmentTotalProps() *Properties {
 		total = CalcProps(*total, e.Hand.Item.Properties)
 	}
 	return total
-}
-
-func GetWeapon(title string, speciality string) *Item {
-	var (
-		weaponProps = new(Properties)
-	)
-	switch speciality {
-	case mage:
-		title += " Sword"
-		weaponProps.Mind = basePoint * 2
-		weaponProps.Mana = baseTen * 2
-	case warrior:
-		title += " staff"
-		weaponProps.Might = basePoint * 2
-		weaponProps.Speed = basePoint * 2
-		weaponProps.Health = baseTen
-	default:
-		title += " branch"
-		weaponProps.Might = basePoint
-	}
-	return &Item{
-		Title:      title,
-		Properties: *weaponProps,
-	}
 }
